// frontend/src/app/models/productAd.model.ts
export class ProductAd {
    nameOfAd: String;
    ownerUsername: String;
    usernameOwner: String;
    typeOfProduct: [String];
    price : Number; 
    structureOfProduct: String;
    weight: Number;
    description: String;
    pictureOfAd: String;
    dateOfPost: String; 
    approvedByAdmin: Boolean;
}