// frontend/src/app/models/accommodationAd.model.ts
export class AccommodationAd {
    nameOfAd: String;
    ownerUsername: String;
    destination: String;
    typeOfAccommodation: [String];
    typeOfService: [String];
    equipment: [String];
    price : Number; 
    description: String;
    pictureOfAd: String;
    dateOfPost: String;
    approvedByAdmin: Boolean;
}